requirements.txt
python-telegram-bot==22.0
requests==2.26.0
Flask==2.2.5
python-dotenv==0.21.1
gunicorn==20.1.0
